Fix issue two identical requests received but one of them is timeout
Sometimes there are two identical requests are sent but one of them
is timeout. BMC should only response the later and ignore the former.
Otherwise host driver will receive duplicate responses which result to
fail to create ssif interfaces.
Below is one of the log demonstrate that scenario
02:23:14 Read ssif request message with len=4 netfn=6 lun=0 cmd=66
02:23:35 Read ssif request message with len=4 netfn=6 lun=0 cmd=66
02:23:47 Send ssif respond message with len=12 netfn=7 lun=0 cmd=66 cc=0
02:23:47 Send ssif respond message with len=12 netfn=7 lun=0 cmd=66 cc=0

This commit fixes this issue by ignoring the first/timeout request

Signed-off-by: quang.name <quangnguyen@os.amperecomputing.com>